backend/requirements.txt
transformers==4.53.3
torch==2.7.1
pandas==2.3.1
requests==2.32.4
numpy>=1.24.0
TA-Lib==0.6.4
openai>=1.0.0
huggingface_hub>=0.20.0
python-dotenv
pytest
pytest-mock
requests-mock
uvicorn
fastapi